requirements.txt
apache-airflow
pandas
boto3
requests
pyarrow
s3fs